refactor : width는 Ladder 객체의 특징
