+++ src/main/java/ladder/domain/game/Ladder.java
@@ -23,9 +23,11 @@ public class Ladder {
     public static final String MESSAGE_LADDER_HEIGHT = "사다리의 높이는 최소 1이상이어야 합니다.";
 
     private final List<Line> ladder;
+    private final int width;
 
     public Ladder(List<Line> ladder) {
         this.ladder = ladder;
+        this.width = ladder.get(0).getLine().size();
     }
 
     public static Ladder of(List<Line> ladder) {
@@ -61,7 +63,7 @@ public class Ladder {
     }
 
     public List<Name> moveAll(Names goals) {
-        return IntStream.range(0, this.ladder.get(0).width())
+        return IntStream.range(0, this.width)
                 .mapToObj(this::move)
                 .map(goals::getParticipantName)
                 .collect(toList());
